# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoAssist Mini API",
    "description": "Service request management API for owners and operators. Protected endpoints require `Authorization: Bearer <firebase-id-token>`. Firebase establishes identity and PostgreSQL establishes application roles.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get the authenticated application user",
        "operationId": "get_me_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tenants": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "List available country tenants",
        "operationId": "list_tenants_api_tenants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantResponse"
                  },
                  "type": "array",
                  "title": "Response List Tenants Api Tenants Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/me/tenant": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Select the authenticated user's country tenant",
        "operationId": "select_tenant_api_me_tenant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "409": {
            "description": "Tenant is already selected or the account is global"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/service-requests": {
      "post": {
        "tags": [
          "service-requests"
        ],
        "summary": "Create a service request",
        "operationId": "create_service_request_api_service_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Owner role required"
          },
          "409": {
            "description": "Country tenant selection is required"
          },
          "422": {
            "description": "Invalid request body"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/service-requests/me": {
      "get": {
        "tags": [
          "service-requests"
        ],
        "summary": "List the authenticated owner's service requests",
        "operationId": "list_my_service_requests_api_service_requests_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceRequestResponse"
                  },
                  "type": "array",
                  "title": "Response List My Service Requests Api Service Requests Me Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Owner role required"
          },
          "409": {
            "description": "Country tenant selection is required"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/operator/service-requests": {
      "get": {
        "tags": [
          "operator"
        ],
        "summary": "List all service requests for operators",
        "operationId": "list_operator_service_requests_api_operator_service_requests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OperatorServiceRequestResponse"
                  },
                  "type": "array",
                  "title": "Response List Operator Service Requests Api Operator Service Requests Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Operator role required"
          },
          "409": {
            "description": "Country tenant selection is required"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/operator/service-requests/{request_id}/assign": {
      "post": {
        "tags": [
          "operator"
        ],
        "summary": "Assign an active workshop to a service request",
        "description": "Assigns a workshop, sets status to ASSIGNED, and publishes service_request.assigned after the database transaction commits. Reassigning an already assigned request updates the workshop and assigned_at and publishes a new event.",
        "operationId": "assign_service_request_api_operator_service_requests__request_id__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignWorkshopRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Workshop is inactive"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Operator role required"
          },
          "409": {
            "description": "Country tenant selection is required"
          },
          "404": {
            "description": "Service request or workshop not found"
          },
          "500": {
            "description": "Assignment persisted but event publication failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workshops": {
      "get": {
        "tags": [
          "workshops"
        ],
        "summary": "List active workshops",
        "operationId": "list_active_workshops_api_workshops_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkshopResponse"
                  },
                  "type": "array",
                  "title": "Response List Active Workshops Api Workshops Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Operator role required"
          },
          "409": {
            "description": "Country tenant selection is required"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List registered users",
        "operationId": "list_users_api_admin_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "type": "array",
                  "title": "Response List Users Api Admin Users Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}/role": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Change a user's OWNER or OPERATOR role",
        "operationId": "update_user_role_api_admin_users__user_id__role_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRoleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}/tenant": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Correct a user's country tenant",
        "operationId": "correct_user_tenant_api_admin_users__user_id__tenant_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantSelection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Administrator account or service-request history prevents correction"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "description": "Verify that the API and database are reachable.",
        "operationId": "health_check_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Check Api Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssignWorkshopRequest": {
        "properties": {
          "workshop_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Workshop Id"
          }
        },
        "type": "object",
        "required": [
          "workshop_id"
        ],
        "title": "AssignWorkshopRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OperatorServiceRequestResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vehicle": {
            "type": "string",
            "title": "Vehicle"
          },
          "problem_type": {
            "$ref": "#/components/schemas/ProblemType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ServiceRequestStatus"
          },
          "assigned_workshop": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkshopSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "assigned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned At"
          },
          "owner": {
            "$ref": "#/components/schemas/OwnerSummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle",
          "problem_type",
          "description",
          "status",
          "assigned_workshop",
          "created_at",
          "assigned_at",
          "owner"
        ],
        "title": "OperatorServiceRequestResponse"
      },
      "OwnerSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "title": "OwnerSummary"
      },
      "ProblemType": {
        "type": "string",
        "enum": [
          "BATTERY",
          "TIRE",
          "MECHANICAL",
          "TOWING",
          "OTHER"
        ],
        "title": "ProblemType"
      },
      "ServiceRequestCreate": {
        "properties": {
          "vehicle": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Vehicle"
          },
          "problem_type": {
            "$ref": "#/components/schemas/ProblemType"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "vehicle",
          "problem_type",
          "description"
        ],
        "title": "ServiceRequestCreate"
      },
      "ServiceRequestResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vehicle": {
            "type": "string",
            "title": "Vehicle"
          },
          "problem_type": {
            "$ref": "#/components/schemas/ProblemType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ServiceRequestStatus"
          },
          "assigned_workshop": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkshopSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "assigned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle",
          "problem_type",
          "description",
          "status",
          "assigned_workshop",
          "created_at",
          "assigned_at"
        ],
        "title": "ServiceRequestResponse"
      },
      "ServiceRequestStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ASSIGNED"
        ],
        "title": "ServiceRequestStatus"
      },
      "TenantResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "TenantResponse"
      },
      "TenantSelection": {
        "properties": {
          "tenant_code": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "pattern": "^[A-Z]{2}$",
            "title": "Tenant Code"
          }
        },
        "type": "object",
        "required": [
          "tenant_code"
        ],
        "title": "TenantSelection"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "firebase_uid": {
            "type": "string",
            "title": "Firebase Uid"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "tenant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "firebase_uid",
          "email",
          "name",
          "role",
          "tenant"
        ],
        "title": "UserResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "OWNER",
          "OPERATOR",
          "ADMIN"
        ],
        "title": "UserRole"
      },
      "UserRoleUpdate": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "OPERATOR"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "UserRoleUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkshopResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "specialty",
          "active"
        ],
        "title": "WorkshopResponse"
      },
      "WorkshopSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "specialty"
        ],
        "title": "WorkshopSummary"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}